ratta: Arc 19 M2 — host data layer (TYPE_DOCUMENT, flags-as-watermark, staleness sweeps)
- SoilSchema.TYPE_DOCUMENT: third additive row type; watermark rides flags
  (epoch millis, NULL = hand-authored); blank-means-absent
- SoilObjectEntity.flags + ClipRow.flags Int? -> Long? (INTEGER affinity
  unchanged — Room identity hash unmoved, proven on a pre-phase file on the
  Nomad); HeadingRows adjusted
- DocumentDao: documentFor + two staleness sweeps + setDocumentText /
  setDocumentDrafted (exactly one moves flags — structural)
- DocumentRepository: get/save/saveDrafted — blank save soft-deletes,
  drop-unchanged-write, watermark moves only in saveDrafted
- Staleness counts soft-deleted rows (og's rule — an erase is a change);
  SN purge wrinkle documented on the query
- liveDescendantIds gains 'document' (page delete/undo/copy carry it)
- NotebookFlags.TEXT_DOCUMENT = 4; all three meta refreshes index-sourced;
  import reads the manifest flag once and feeds both writes
- 1161 JVM tests/variant (+35); purge/clipboard/undo parity pinned

+++ apps/notesprout_ratta/app/src/main/kotlin/com/symmetricalpalmtree/notesproutsn/data/clip/ClipEnvelope.kt
@@ -28,7 +28,9 @@ data class ClipRow(
     val color: String? = null,
     val strokeWidth: Float? = null,
     val style: String? = null,
-    val flags: Int? = null,
+    /** `Long` since arc 19, mirroring the entity: a `document` row's watermark is epoch millis.
+     *  JSON-compatible both ways — an older envelope's small ints decode as longs unchanged. */
+    val flags: Long? = null,
     val blob: String? = null,
 ) {
     /** The decoded blob, or null when there is none **or it is unusable** — a malformed stroke
